Refactor : NotificationModal이 toast의 기능을 담당하는 것 갈아 toast로 수정
